Completed BALAD with full debug functionality
Developed a fully functioning Reverse Polish Notation calculator
'george.bl' in the BALAD assembler language. This calculator has add,
subtract, multiply and division operators as well as negate. Numbers
are entered and pushed on a stack. The operators act on the last
two numbers on the stack, which is the way RPN calculators work.
Additional functions are Clear the stack, Enter or duplicate the last
number on the stack, Drop the last number from the stack, Swap the last
two numbers on the stack and Roll the last number with the number at
the top of the stack. Numbers are integers only and can be entered in
decimal, octal with a leading 0 as well as hexadecimal and binary with
leading 0x and 0b. These conversion are done in BALAD code. Results
can be shown in decimal, octal, hexadecimal or binary.  All this had
to fit into 512 words of BALAD memory, which was a challenge.

Historical Note: I wrote my first computer program in 1957 for a
Reverse Polish Notation Calculater called 'GEORGE', which ran as an
interpreter on the 'UTECOM' computer at the University of NSW. UTECOM
was the third digital electronic computer installed in Australia. It
nearly filled a large room in the first building on the campus in
Kensington.

GEORGE had been written by one of my lecturers, who was teaching us
Boolean Logic as part of a philosophy course. In 1956 UTECOM had
384 x 32 bit words of memory implemented as 12 mercury delay lines.
This is very similar in memory size and precision to the BALAD
computer, which was first implemented by me in 1970 as a teaching aid.

I have named this program 'george.bl' in honour of my former lecturer.

Also developed three programs - pixel.bl, smiley.bl and felix.bl -
to demonstrate computer graphics in BALAD using character graphics.
This shows the advances in the use of computers - mainly for number
crunching in the 50's to the addition of text processing in the 60's
to pixel graphics in the 80's.

balad:        Fixed some bugs. Added many features. Tracing is now
	      done with symbolic code where possible and the jump
	      tester output has been added. Watch point tracing is
	      marked with the word 'watch'.  Extensive testing was
	      done in conjunction with the development of george.bl
enterCH.bl:   small changes
felix.bl:     a smiley with a frame dedicated to my grandson Felix, who
              inspired me to start this project
george.bl:    new program using the full power of the BALAD language
infinite.bl:  more functionality
interrupt.bl: small changes
pixel.bl:     character graphics using the bitmaps from 'immediate C'
smiley.bl:    for your enjoyment

    Id: BALAD2020/balad 1.4
    Id: BALAD2020/enterCH.bl 1.2
    Id: BALAD2020/felix.bl 1.1
    Id: BALAD2020/george.bl 1.1
    Id: BALAD2020/infinite.bl 1.2
    Id: BALAD2020/interrupt.bl 1.2
    Id: BALAD2020/pixel.bl 1.1
    Id: BALAD2020/smiley.bl 1.1